refactor: create ColumnDefinitions
